// src/utils/types.ts
export interface BeersType {
    id:          string;
    name:        string;
    img:         string;
    category:    string;
    color:       string;
    origin:      string;
    beerType:    string;
    beerTone:    string;
    alcohol:     string;
    description: string;
    url:         string;
}